requirements.txt

requests==2.32.3
lxml==5.2.2
cssselect==1.2.0
PyYAML==6.0.2
fastapi==0.115.0
uvicorn==0.30.6
python-dateutil==2.9.0.post0
pillow==10.4.0
ImageHash==4.3.1
charset-normalizer==3.3.2
idna==3.7
urllib3==2.2.3
certifi==2024.8.30
